dashboard/requirements.txt
streamlit>=1.30.0
psycopg2-binary>=2.9.9
pandas>=2.1.0
plotly>=5.18.0
